requirements.txt
requests>=2.31.0
websocket-client>=1.7.0
schedule>=1.2.0
python-telegram-bot>=20.8
python-dotenv>=1.0.0
asyncio>=3.4.3
aiohttp>=3.9.0